Support for custom headers
This generally allows any header to be sent. Supplied headers are merged
with the default headers and can overwrite them.

The specific purpose of this is to allow the `X-Request-ID` header to be
specified. The request ID will be set on the Cube server then too, so
requests can be traced across services.